+++ SentryTestUtils/TestCurrentDateProvider.swift
@@ -2,64 +2,103 @@ import Foundation
 @testable import Sentry
 
 public class TestCurrentDateProvider: SentryCurrentDateProvider {
+
     public static let defaultStartingDate = Date(timeIntervalSinceReferenceDate: 0)
+
     private var internalDate = defaultStartingDate
     private var internalSystemTime: UInt64 = 0
+
     public var driftTimeForEveryRead = false
     public var driftTimeInterval = 0.1
+
     private var _systemUptime: TimeInterval = 0
+
+    // NSLock isn't reentrant, so we use NSRecursiveLock.
+    private let lock = NSRecursiveLock()
 
     public init() {
     }
 
     public func date() -> Date {
-        defer {
-            if driftTimeForEveryRead {
-                internalDate = internalDate.addingTimeInterval(driftTimeInterval)
+        return lock.synchronized {
+            defer {
+                if driftTimeForEveryRead {
+                    internalDate = internalDate.addingTimeInterval(driftTimeInterval)
+                }
             }
-        }
 
-        return internalDate
+            return internalDate
+        }
     }
 
     /// Reset the date provider to its default starting date.
     public func reset() {
-        setDate(date: TestCurrentDateProvider.defaultStartingDate)
-        internalSystemTime = 0
+        lock.synchronized {
+            setDate(date: TestCurrentDateProvider.defaultStartingDate)
+            internalSystemTime = 0
+        }
     }
 
     public func setDate(date: Date) {
-        internalDate = date
+        lock.synchronized {
+            internalDate = date
+        }
     }
 
     /// Advance the current date by the specified number of seconds.
     public func advance(by seconds: TimeInterval) {
-        setDate(date: date().addingTimeInterval(seconds))
-        internalSystemTime += seconds.toNanoSeconds()
+        lock.synchronized {
+            setDate(date: date().addingTimeInterval(seconds))
+            internalSystemTime += seconds.toNanoSeconds()
+        }
     }
 
     public func advanceBy(nanoseconds: UInt64) {
-        setDate(date: date().addingTimeInterval(nanoseconds.toTimeInterval()))
-        internalSystemTime += nanoseconds
+        lock.synchronized {
+            setDate(date: date().addingTimeInterval(nanoseconds.toTimeInterval()))
+            internalSystemTime += nanoseconds
+        }
     }
 
     public func advanceBy(interval: TimeInterval) {
-        setDate(date: date().addingTimeInterval(interval))
-        internalSystemTime += interval.toNanoSeconds()
-    }
-
-    public var timezoneOffsetValue = 0
-    public func timezoneOffset() -> Int {
-        return timezoneOffsetValue
+        lock.synchronized {
+            setDate(date: date().addingTimeInterval(interval))
+            internalSystemTime += interval.toNanoSeconds()
+        }
     }
 
     public func systemTime() -> UInt64 {
-        return internalSystemTime
+        lock.synchronized {
+            return internalSystemTime
+        }
     }
     public func systemUptime() -> TimeInterval {
-        _systemUptime
+        return lock.synchronized {
+            return _systemUptime
+        }
     }
     public func setSystemUptime(_ uptime: TimeInterval) {
-        _systemUptime = uptime
+        lock.synchronized {
+            _systemUptime = uptime
+        }
+    }
+
+    private var _timezoneOffsetValue = 0
+    public var timezoneOffsetValue: Int {
+        get {
+            return lock.synchronized {
+                return _timezoneOffsetValue
+            }
+        }
+        set {
+            lock.synchronized {
+                _timezoneOffsetValue = newValue
+            }
+        }
+    }
+    public func timezoneOffset() -> Int {
+        return lock.synchronized {
+            return _timezoneOffsetValue
+        }
     }
 }