+++ Tests/SentryProfilerTests/SentrySamplingProfilerTests.mm
@@ -9,6 +9,7 @@
 #    import "SentryThreadMetadataCache.hpp"
 #    import "SentryTime.h"
 
+#    import "SentrySwift.h"
 #    import <chrono>
 #    import <iostream>
 #    import <pthread.h>
@@ -41,7 +42,8 @@ using namespace sentry::profiling;
     XCTAssertFalse(profiler->isSampling());
 
     std::uint64_t start = 0;
-    profiler->startSampling([&start] { start = getAbsoluteTime(); });
+    profiler->startSampling(
+        [&start] { start = [SentryDefaultCurrentDateProvider getAbsoluteTime]; });
     XCTAssertTrue(profiler->isSampling());
 
     // sleep long enough for 2 samples to be collected
@@ -53,7 +55,8 @@ using namespace sentry::profiling;
     XCTAssertFalse(profiler->isSampling());
 
     XCTAssertGreaterThan(start, static_cast<std::uint64_t>(0));
-    XCTAssertGreaterThan(getDurationNs(start, getAbsoluteTime()), 0ULL);
+    XCTAssertGreaterThan(
+        getDurationNs(start, [SentryDefaultCurrentDateProvider getAbsoluteTime]), 0ULL);
     XCTAssertGreaterThan(profiler->numSamples(), static_cast<std::uint64_t>(0));
     XCTAssertGreaterThan(numIdleSamples, 0);
 }