+++ Tests/SentryTests/Integrations/ANR/SentryANRTrackerV1Tests.swift
@@ -1,5 +1,5 @@
-@testable import Sentry
-import SentryTestUtils
+@_spi(Private) @testable import Sentry
+@_spi(Private) import SentryTestUtils
 import XCTest
 
 #if os(iOS) || os(tvOS) || targetEnvironment(macCatalyst)
@@ -115,8 +115,6 @@ class SentryANRTrackerV1Tests: XCTestCase, SentryANRTrackerDelegate {
     }
     
     func testAppSuspended_NoANR() {
-        // To avoid spamming the test logs
-        SentryLog.configure(true, diagnosticLevel: .error)
         
         anrDetectedExpectation.isInverted = true
         fixture.dispatchQueue.blockBeforeMainBlock = {
@@ -127,8 +125,6 @@ class SentryANRTrackerV1Tests: XCTestCase, SentryANRTrackerDelegate {
         start()
         
         wait(for: [anrDetectedExpectation, anrStoppedExpectation], timeout: waitTimeout)
-        
-        SentryLog.setTestDefaultLogLevel()
     }
     
     func testRemoveListener_StopsReportingANRs() {