+++ Tests/SentryTests/Integrations/SessionReplay/SentryOnDemandReplayTests.swift
@@ -8,9 +8,8 @@ import XCTest
 #if os(iOS) || os(tvOS)
 class SentryOnDemandReplayTests: XCTestCase {
     
-    private let dateProvider = TestCurrentDateProvider()
     private var outputPath = FileManager.default.temporaryDirectory.appendingPathComponent("replayTest")
-    
+
     override func setUpWithError() throws {
         try removeDirectoryIfExists(at: outputPath)
         try FileManager.default.createDirectory(at: outputPath, withIntermediateDirectories: true)
@@ -28,119 +27,130 @@ class SentryOnDemandReplayTests: XCTestCase {
     }
     
     private func getSut(trueDispatchQueueWrapper: Bool = false) -> SentryOnDemandReplay {
-        return SentryOnDemandReplay(outputPath: outputPath.path,
-                                    workingQueue: trueDispatchQueueWrapper ? SentryDispatchQueueWrapper() : TestSentryDispatchQueueWrapper(),
-                                    dateProvider: dateProvider)
+        return SentryOnDemandReplay(
+            outputPath: outputPath.path,
+            processingQueue: trueDispatchQueueWrapper ? SentryDispatchQueueWrapper() : TestSentryDispatchQueueWrapper(),
+            assetWorkerQueue: trueDispatchQueueWrapper ? SentryDispatchQueueWrapper() : TestSentryDispatchQueueWrapper()
+        )
     }
     
     func testAddFrame() {
         let sut = getSut()
-        sut.addFrameAsync(image: UIImage.add)
-       
+        sut.addFrameAsync(timestamp: Date(timeIntervalSinceReferenceDate: 0xBAAD_F00D), maskedViewImage: UIImage.add)
+
         guard let frame = sut.frames.first else {
             XCTFail("Frame was not saved")
             return
         }
-        
+
         XCTAssertEqual(FileManager.default.fileExists(atPath: frame.imagePath), true)
         XCTAssertEqual(frame.imagePath.hasPrefix(self.outputPath.path), true)
     }
     
     func testReleaseFrames() {
         let sut = getSut()
-        
-        for _ in 0..<10 {
-            sut.addFrameAsync(image: UIImage.add)
-            dateProvider.advance(by: 1)
+
+        let start = Date(timeIntervalSinceReferenceDate: 0)
+        for i in 0..<10 {
+            sut.addFrameAsync(timestamp: start.addingTimeInterval(TimeInterval(i)), maskedViewImage: UIImage.add)
         }
+        let end = start.addingTimeInterval(10)
        
-        sut.releaseFramesUntil(dateProvider.date().addingTimeInterval(-5))
-        
+        sut.releaseFramesUntil(end.addingTimeInterval(-5))
+
         let frames = sut.frames
-        
+
         XCTAssertEqual(frames.count, 5)
-        XCTAssertEqual(frames.first?.time, Date(timeIntervalSinceReferenceDate: 5))
-        XCTAssertEqual(frames.last?.time, Date(timeIntervalSinceReferenceDate: 9))
+        XCTAssertEqual(frames.first?.time, start.addingTimeInterval(5))
+        XCTAssertEqual(frames.last?.time, start.addingTimeInterval(9))
     }
     
     func testFramesWithScreenName() {
         let sut = getSut()
-        
+
+        let start = Date(timeIntervalSinceReferenceDate: 0)
         for i in 0..<4 {
-            sut.addFrameAsync(image: UIImage.add, forScreen: "\(i)")
-            dateProvider.advance(by: 1)
+            sut.addFrameAsync(timestamp: start.addingTimeInterval(TimeInterval(i)), maskedViewImage: UIImage.add, forScreen: "\(i)")
         }
-        
-        sut.releaseFramesUntil(dateProvider.date().addingTimeInterval(-5))
-        
+
+        sut.releaseFramesUntil(start)
+
         let frames = sut.frames
-        
+
         for i in 0..<4 {
             XCTAssertEqual(frames[i].screenName, "\(i)")
         }
     }
     
     func testGenerateVideo() throws {
+        // -- Arrange --
         let sut = getSut()
-        dateProvider.driftTimeForEveryRead = true
-        dateProvider.driftTimeInterval = 1
-        
-        for _ in 0..<10 {
-            sut.addFrameAsync(image: UIImage.add)
+
+        let start = Date(timeIntervalSinceReferenceDate: 0)
+        for i in 0..<10 {
+            sut.addFrameAsync(timestamp: start.addingTimeInterval(TimeInterval(i)), maskedViewImage: UIImage.add)
         }
-        
+        let end = start.addingTimeInterval(10)
+
         let videoExpectation = expectation(description: "Wait for video render")
-        
-        let videos = try sut.createVideoWith(beginning: Date(timeIntervalSinceReferenceDate: 0), end: Date(timeIntervalSinceReferenceDate: 10))
+
+        // -- Act --
+        let videos = sut.createVideoWith(beginning: start, end: end)
+
+        // -- Assert --
         XCTAssertEqual(videos.count, 1)
         let info = try XCTUnwrap(videos.first)
-        
+
         XCTAssertEqual(info.duration, 10)
         XCTAssertEqual(info.start, Date(timeIntervalSinceReferenceDate: 0))
         XCTAssertEqual(info.end, Date(timeIntervalSinceReferenceDate: 10))
-        
+
         let videoPath = info.path
-        
+
         XCTAssertTrue(FileManager.default.fileExists(atPath: videoPath.path))
-        
+
         videoExpectation.fulfill()
         try FileManager.default.removeItem(at: videoPath)
         wait(for: [videoExpectation], timeout: 1)
     }
     
     func testAddFrameIsThreadSafe() {
-        let queue = SentryDispatchQueueWrapper()
-        let sut = SentryOnDemandReplay(outputPath: outputPath.path,
-                                       workingQueue: queue,
-                                       dateProvider: dateProvider)
-        
-        dateProvider.driftTimeForEveryRead = true
-        dateProvider.driftTimeInterval = 1
+        let processingQueue = SentryDispatchQueueWrapper()
+        let workerQueue = SentryDispatchQueueWrapper()
+        let sut = SentryOnDemandReplay(
+            outputPath: outputPath.path,
+            processingQueue: processingQueue,
+            assetWorkerQueue: workerQueue
+        )
         let group = DispatchGroup()
-        
-        for _ in 0..<10 {
+
+        let start = Date(timeIntervalSinceReferenceDate: 0)
+        for i in 0..<10 {
             group.enter()
             DispatchQueue.global().async {
-                sut.addFrameAsync(image: UIImage.add)
+                sut.addFrameAsync(timestamp: start.addingTimeInterval(TimeInterval(i)), maskedViewImage: UIImage.add)
                 group.leave()
             }
         }
-        
+
         group.wait()
-        queue.queue.sync {} //Wait for all enqueued operation to finish
+        processingQueue.queue.sync {} // Wait for all enqueued operation to finish
         XCTAssertEqual(sut.frames.map({ ($0.imagePath as NSString).lastPathComponent }), (0..<10).map { "\($0).0.png" })
     }
     
     func testReleaseIsThreadSafe() {
-        let queue = SentryDispatchQueueWrapper()
-        let sut = SentryOnDemandReplay(outputPath: outputPath.path,
-                                       workingQueue: queue,
-                                       dateProvider: dateProvider)
-                
+        let processingQueue = SentryDispatchQueueWrapper()
+        let workerQueue = SentryDispatchQueueWrapper()
+        let sut = SentryOnDemandReplay(
+            outputPath: outputPath.path,
+            processingQueue: processingQueue,
+            assetWorkerQueue: workerQueue
+        )
+
         sut.frames = (0..<100).map { SentryReplayFrame(imagePath: outputPath.path + "/\($0).png", time: Date(timeIntervalSinceReferenceDate: Double($0)), screenName: nil) }
-                
+
         let group = DispatchGroup()
-        
+
         for i in 1...10 {
             group.enter()
             DispatchQueue.global().async {
@@ -148,61 +158,75 @@ class SentryOnDemandReplayTests: XCTestCase {
                 group.leave()
             }
         }
-        
+
         group.wait()
-        
-        queue.queue.sync {} //Wait for all enqueued operation to finish
+
+        processingQueue.queue.sync {} //Wait for all enqueued operation to finish
         XCTAssertEqual(sut.frames.count, 0)
     }
     
-    func testInvalidWriter() throws {
-        let queue = TestSentryDispatchQueueWrapper()
-        let sut = SentryOnDemandReplay(outputPath: outputPath.path,
-                                       workingQueue: queue,
-                                       dateProvider: dateProvider)
-        
-        let start = dateProvider.date()
-        sut.addFrameAsync(image: UIImage.add)
-        dateProvider.advance(by: 1)
-        let end = dateProvider.date()
-        
-        //Creating a file where the replay would be written to cause an error in the writer
-        try Data("tempFile".utf8).write(to: outputPath.appendingPathComponent("0.0.mp4"))
-        
-        XCTAssertThrowsError(try sut.createVideoWith(beginning: start, end: end))
+    func testCreateVideo_outputFileAlreadyExists_assetWriterErrorShouldNotThrowError() throws {
+        // -- Arrange --
+        let processingQueue = SentryDispatchQueueWrapper()
+        let workerQueue = SentryDispatchQueueWrapper()
+        let sut = SentryOnDemandReplay(
+            outputPath: outputPath.path,
+            processingQueue: processingQueue,
+            assetWorkerQueue: workerQueue
+        )
+
+        let start = Date(timeIntervalSinceReferenceDate: 0xBAAD_F00D)
+        sut.addFrameAsync(timestamp: start, maskedViewImage: UIImage.add)
+        processingQueue.dispatchSync {
+            // Wait for the frame to be added by adding a sync operation to the serial queue
+        }
+        let end = start.addingTimeInterval(1)
+
+        // Creating a file where the replay would be written to cause an error in the writer
+        let expectedOutputPath = outputPath
+            .appendingPathComponent("\(start.timeIntervalSinceReferenceDate)")
+            .appendingPathExtension("mp4")
+        try Data("tempFile".utf8).write(to: expectedOutputPath)
+
+        // -- Act & Assert --
+        let result = sut.createVideoWith(beginning: start, end: end)
+        XCTAssertEqual(result.count, 0)
     }
     
     func testGenerateVideoForEachSize() throws {
+        // -- Arrange --
         let sut = getSut()
-        dateProvider.driftTimeForEveryRead = true
-        dateProvider.driftTimeInterval = 1
-        
+
         let image1 = UIGraphicsImageRenderer(size: CGSize(width: 20, height: 19)).image { _ in }
         let image2 = UIGraphicsImageRenderer(size: CGSize(width: 20, height: 10)).image { _ in }
-        
+
+        let start = Date(timeIntervalSinceReferenceDate: 0)
         for i in 0..<10 {
-            sut.addFrameAsync(image: i < 5 ? image1 : image2)
+            sut.addFrameAsync(timestamp: start.addingTimeInterval(TimeInterval(i)), maskedViewImage: i < 5 ? image1 : image2)
         }
-        
-        let videos = try sut.createVideoWith(beginning: Date(timeIntervalSinceReferenceDate: 0), end: Date(timeIntervalSinceReferenceDate: 10))
-        
+        let end = start.addingTimeInterval(10)
+
+        // -- Act --
+        let videos = sut.createVideoWith(beginning: start, end: end)
+
+        // -- Assert --
         XCTAssertEqual(videos.count, 2)
-        
+
         let firstVideo = try XCTUnwrap(videos.first)
         let secondVideo = try XCTUnwrap(videos.last)
-        
+
         XCTAssertEqual(firstVideo.duration, 5)
         XCTAssertEqual(secondVideo.duration, 5)
-        
-        XCTAssertEqual(firstVideo.start, Date(timeIntervalSinceReferenceDate: 0))
-        XCTAssertEqual(secondVideo.start, Date(timeIntervalSinceReferenceDate: 5))
-        
-        XCTAssertEqual(firstVideo.end, Date(timeIntervalSinceReferenceDate: 5))
-        XCTAssertEqual(secondVideo.end, Date(timeIntervalSinceReferenceDate: 10))
-        
+
+        XCTAssertEqual(firstVideo.start, start)
+        XCTAssertEqual(secondVideo.start, start.addingTimeInterval(5))
+
+        XCTAssertEqual(firstVideo.end, start.addingTimeInterval(5))
+        XCTAssertEqual(secondVideo.end, end)
+
         XCTAssertEqual(firstVideo.width, 20)
         XCTAssertEqual(firstVideo.height, 19)
-        
+
         XCTAssertEqual(secondVideo.width, 20)
         XCTAssertEqual(secondVideo.height, 10)
     }
@@ -210,19 +234,16 @@ class SentryOnDemandReplayTests: XCTestCase {
     func testGenerateVideoInfo_whenNoFramesAdded_shouldNotThrowError() throws {
         // -- Arrange --
         let sut = getSut()
-        dateProvider.driftTimeForEveryRead = true
-        dateProvider.driftTimeInterval = 1
-
         // -- Act --
-        let videos = try sut.createVideoWith(
+        let videos = sut.createVideoWith(
             beginning: Date(timeIntervalSinceReferenceDate: 0),
             end: Date(timeIntervalSinceReferenceDate: 10)
         )
 
         // -- Assert --
-        XCTAssertNil(videos.first)
+        XCTAssertEqual(videos.count, 0)
     }
-  
+    
     func testCalculatePresentationTime_withOneFPS_shouldReturnTiming() {
         // -- Arrange --
         let framesPerSecond = 1
@@ -317,7 +338,7 @@ class SentryOnDemandReplayTests: XCTestCase {
         XCTAssertEqual(settings[AVVideoCodecKey] as? AVVideoCodecType, AVVideoCodecType.h264)
         XCTAssertEqual(settings[AVVideoWidthKey] as? CGFloat, 0)
         XCTAssertEqual(settings[AVVideoHeightKey] as? CGFloat, 0)
-        
+
         let compressionProperties = try XCTUnwrap(settings[AVVideoCompressionPropertiesKey] as? [String: Any], "Compression properties not found")
 
         XCTAssertEqual(compressionProperties.count, 4)
@@ -325,7 +346,7 @@ class SentryOnDemandReplayTests: XCTestCase {
         XCTAssertEqual(compressionProperties[AVVideoProfileLevelKey] as? String, AVVideoProfileLevelH264MainAutoLevel)
         XCTAssertEqual(compressionProperties[AVVideoAllowFrameReorderingKey] as? Bool, false)
         XCTAssertEqual(compressionProperties[AVVideoMaxKeyFrameIntervalKey] as? Int, 6)
-        
+
         let colorProperties = try XCTUnwrap(settings[AVVideoColorPropertiesKey] as? [String: Any], "Color properties not found")
 
         XCTAssertEqual(colorProperties.count, 3)
@@ -333,7 +354,7 @@ class SentryOnDemandReplayTests: XCTestCase {
         XCTAssertEqual(colorProperties[AVVideoTransferFunctionKey] as? String, AVVideoTransferFunction_ITU_R_709_2)
         XCTAssertEqual(colorProperties[AVVideoYCbCrMatrixKey] as? String, AVVideoYCbCrMatrix_ITU_R_709_2)
     }
-
+    
     func testCreateVideoSettings_anySize_shouldReturnFullSettings() throws {
         // -- Arrange --
         let sut = getSut()
@@ -346,7 +367,7 @@ class SentryOnDemandReplayTests: XCTestCase {
         XCTAssertEqual(settings[AVVideoCodecKey] as? AVVideoCodecType, AVVideoCodecType.h264)
         XCTAssertEqual(settings[AVVideoWidthKey] as? CGFloat, 100)
         XCTAssertEqual(settings[AVVideoHeightKey] as? CGFloat, 100)
-        
+
         let compressionProperties = try XCTUnwrap(settings[AVVideoCompressionPropertiesKey] as? [String: Any], "Compression properties not found")
 
         XCTAssertEqual(compressionProperties.count, 4)
@@ -354,7 +375,7 @@ class SentryOnDemandReplayTests: XCTestCase {
         XCTAssertEqual(compressionProperties[AVVideoProfileLevelKey] as? String, AVVideoProfileLevelH264MainAutoLevel)
         XCTAssertEqual(compressionProperties[AVVideoAllowFrameReorderingKey] as? Bool, false)
         XCTAssertEqual(compressionProperties[AVVideoMaxKeyFrameIntervalKey] as? Int, 6)
-        
+
         let colorProperties = try XCTUnwrap(settings[AVVideoColorPropertiesKey] as? [String: Any], "Color properties not found")
 
         XCTAssertEqual(colorProperties.count, 3)