+++ Tests/SentryTests/TestUtils/TestConncurrentModifications.swift
@@ -1,9 +1,8 @@
 import Foundation
-import SentryTestUtils
+@_spi(Private) import Sentry
+@_spi(Private) import SentryTestUtils
 
 func testConcurrentModifications(asyncWorkItems: Int = 5, writeLoopCount: Int = 1_000, writeWork: @escaping (Int) -> Void, readWork: @escaping () -> Void = {}) {
-    // To not spam the test logs
-    SentryLog.configureLog(true, diagnosticLevel: .error)
     
     let queue = DispatchQueue(label: "testConcurrentModifications", qos: .userInteractive, attributes: [.concurrent, .initiallyInactive])
     let group = DispatchGroup()
@@ -24,6 +23,4 @@ func testConcurrentModifications(asyncWorkItems: Int = 5, writeLoopCount: Int = 
     
     queue.activate()
     group.waitWithTimeout(timeout: 500)
-    
-    SentryLog.setTestDefaultLogLevel()
 }